refactor(MonitorGraph): splits monitor data control into it's own component
